# app_bi/api/swag/swagger.yaml
definitions:
...
  github_com_leaf-rain_raindata_app_bi_internal_data_data.SysAuthority:
    properties:
      authorityId:
        description: 角色ID
        type: integer
      authorityName:
        description: 角色名
        type: string
      children:
        items:
          $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_internal_data_data.SysAuthority'
        type: array
      createdAt:
        description: 创建时间
        type: string
      dataAuthorityId:
        items:
          $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_internal_data_data.SysAuthority'
        type: array
      defaultRouter:
        description: 默认菜单(默认dashboard)
        type: string
      deletedAt:
        type: string
      parentId:
        description: 父角色ID
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
...
  github_com_leaf-rain_raindata_app_bi_internal_data_data.SysUser:
    properties:
      authorities:
        items:
          $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_internal_data_data.SysAuthority'
        type: array
      authority:
        $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_internal_data_data.SysAuthority'
      authorityId:
        description: 用户角色ID
        type: integer
      baseColor:
        description: 基础颜色
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        description: 用户邮箱
        type: string
      enable:
        description: 用户是否被冻结 1正常 2冻结
        type: integer
      headerImg:
        description: 用户头像
        type: string
      id:
        type: integer
      nickName:
        description: 用户昵称
        type: string
      password:
        description: 用户登录密码
        type: string
      phone:
        description: 用户手机号
        type: string
      sideMode:
        description: 用户侧边主题
        type: string
      updatedAt:
        type: string
      userName:
        description: 用户登录名
        type: string
    type: object
...
  github_com_leaf-rain_raindata_app_bi_internal_data_dto.ChangePasswordReq:
    properties:
      newPassword:
        description: 新密码
        type: string
      password:
        description: 密码
        type: string
    type: object
  github_com_leaf-rain_raindata_app_bi_internal_data_dto.Login:
    properties:
      captcha:
        description: 验证码
        type: string
      captchaId:
        description: 验证码ID
        type: string
      password:
        description: 密码
        type: string
      username:
        description: 用户名
        type: string
    type: object
  github_com_leaf-rain_raindata_app_bi_internal_data_dto.LoginResponse:
    properties:
      expiresAt:
        type: integer
      token:
        type: string
      user:
        $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_internal_data_data.SysUser'
    type: object
...
  github_com_leaf-rain_raindata_app_bi_internal_data_dto.Register:
    properties:
      authorityId:
        example: int 角色id
        type: string
      authorityIds:
        example: '[]uint 角色id'
        type: string
      email:
        example: 电子邮箱
        type: string
      enable:
        example: int 是否启用
        type: string
      headerImg:
        example: 头像链接
        type: string
      nickName:
        example: 昵称
        type: string
      passWord:
        example: 密码
        type: string
      phone:
        example: 电话号码
        type: string
      userName:
        example: 用户名
        type: string
    type: object
  github_com_leaf-rain_raindata_app_bi_internal_data_dto.SetUserAuth:
    properties:
      authorityId:
        description: 角色ID
        type: integer
    type: object
  github_com_leaf-rain_raindata_app_bi_internal_data_dto.SetUserAuthorities:
    properties:
      authorityIds:
        description: 角色ID
        items:
          type: integer
        type: array
      id:
        type: integer
    type: object
  github_com_leaf-rain_raindata_app_bi_internal_data_dto.SysUserResponse:
    properties:
      user:
        $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_internal_data_data.SysUser'
    type: object
  github_com_leaf-rain_raindata_app_bi_third_party_rhttp.GetById:
    properties:
      id:
        description: 主键ID
        type: integer
    type: object
...
  github_com_leaf-rain_raindata_app_bi_third_party_rhttp.PageInfo:
    properties:
      keyword:
        description: 关键字
        type: string
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
    type: object
  github_com_leaf-rain_raindata_app_bi_third_party_rhttp.PageResult:
    properties:
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  github_com_leaf-rain_raindata_app_bi_third_party_rhttp.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
info:
  contact: {}
paths:
...
  /base/login:
    post:
      parameters:
      - description: 用户名, 密码, 验证码
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_internal_data_dto.Login'
      produces:
      - application/json
      responses:
        "200":
          description: 返回包括用户信息,token,过期时间
          schema:
            allOf:
            - $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_third_party_rhttp.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_internal_data_dto.LoginResponse'
                msg:
                  type: string
              type: object
      summary: 用户登录
      tags:
      - Base
...
  /user/SetSelfInfo:
    put:
      consumes:
      - application/json
      parameters:
      - description: ID, 用户名, 昵称, 头像链接
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_internal_data_data.SysUser'
      produces:
      - application/json
      responses:
        "200":
          description: 设置用户信息
          schema:
            allOf:
            - $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_third_party_rhttp.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 设置用户信息
      tags:
      - SysUser
  /user/admin_register:
    post:
      parameters:
      - description: 用户名, 昵称, 密码, 角色ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_internal_data_dto.Register'
      produces:
      - application/json
      responses:
        "200":
          description: 用户注册账号,返回包括用户信息
          schema:
            allOf:
            - $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_third_party_rhttp.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_internal_data_dto.SysUserResponse'
                msg:
                  type: string
              type: object
      summary: 用户注册账号
      tags:
      - SysUser
  /user/changePassword:
    post:
      parameters:
      - description: 用户名, 原密码, 新密码
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_internal_data_dto.ChangePasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: 用户修改密码
          schema:
            allOf:
            - $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_third_party_rhttp.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 用户修改密码
      tags:
      - SysUser
  /user/deleteUser:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_third_party_rhttp.GetById'
      produces:
      - application/json
      responses:
        "200":
          description: 删除用户
          schema:
            allOf:
            - $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_third_party_rhttp.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除用户
      tags:
      - SysUser
  /user/getUserInfo:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 获取用户信息
          schema:
            allOf:
            - $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_third_party_rhttp.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取用户信息
      tags:
      - SysUser
  /user/getUserList:
    post:
      consumes:
      - application/json
      parameters:
      - description: 页码, 每页大小
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_third_party_rhttp.PageInfo'
      produces:
      - application/json
      responses:
        "200":
          description: 分页获取用户列表,返回包括列表,总数,页码,每页数量
          schema:
            allOf:
            - $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_third_party_rhttp.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_third_party_rhttp.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取用户列表
      tags:
      - SysUser
  /user/resetPassword:
    post:
      parameters:
      - description: ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_internal_data_data.SysUser'
      produces:
      - application/json
      responses:
        "200":
          description: 重置用户密码
          schema:
            allOf:
            - $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_third_party_rhttp.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 重置用户密码
      tags:
      - SysUser
  /user/setUserAuthorities:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户UUID, 角色ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_internal_data_dto.SetUserAuthorities'
      produces:
      - application/json
      responses:
        "200":
          description: 设置用户权限
          schema:
            allOf:
            - $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_third_party_rhttp.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 设置用户权限
      tags:
      - SysUser
  /user/setUserAuthority:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户UUID, 角色ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_internal_data_dto.SetUserAuth'
      produces:
      - application/json
      responses:
        "200":
          description: 设置用户权限
          schema:
            allOf:
            - $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_third_party_rhttp.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 更改用户权限
      tags:
      - SysUser
  /user/setUserInfo:
    put:
      consumes:
      - application/json
      parameters:
      - description: ID, 用户名, 昵称, 头像链接
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_internal_data_data.SysUser'
      produces:
      - application/json
      responses:
        "200":
          description: 设置用户信息
          schema:
            allOf:
            - $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_third_party_rhttp.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 设置用户信息
      tags:
      - SysUser
swagger: "2.0"